+++ src/lib/harness/meshVertex.glsl
@@ -5,10 +5,13 @@ uniform mat4 modelViewMatrix;
 
 attribute vec3 position;
 attribute vec3 normal;
+attribute vec2 uv;
 
 varying vec3 vNormal;
+varying vec2 vUv;
 
 void main() {
     vNormal = normal;
+    vUv = uv;
     gl_Position = projectionMatrix * modelViewMatrix * vec4(position, 1.0);
 }